DT-657: Fixes #3785: Split up examples:init command, don't run automatically.
