+++ src/Robo/Commands/Blt/UpdateCommand.php
@@ -113,7 +113,6 @@ class UpdateCommand extends BltTasks {
     $this->getConfig()->replace($new_config->export());
 
     $this->invokeCommand('blt:init:settings');
-    $this->invokeCommand('recipes:blt:init:command');
     $this->invokeCommand('blt:init:shell-alias');
     if (DIRECTORY_SEPARATOR === '\\') {
       // On Windows, during composer create-project,
